requirements.txt
h11~=0.14.0
pip~=24.0
logging~=0.4.9.6
typing_extensions~=4.12.2
sniffio~=1.3.1
httpcore~=1.0.7
idna~=3.10
httpx~=0.28.1
certifi~=2024.12.14
anyio~=4.8.0
python-telegram-bot~=21.10